shell - Update sync script
- Move only i3 files when they are existing on the current machine
- REmove some no longer needed files from the exclude list

+++ sync.sh
@@ -3,11 +3,13 @@ cd "$(dirname "$0")"
 git pull
 
 copyFiles() {
-  rsync --exclude ".git/" --exclude ".DS_Store" --exclude "sync.sh" --exclude "README.md" --exclude "iTerm2" --exclude "terminal" --exclude "install-deps.sh" --exclude "readme.md" --exclude "createlinks.sh" -av . ~
+  rsync --exclude ".git/" --exclude --exclude "sync.sh" --exclude "README.md" --exclude "terminal" --exclude "readme.md" -av . ~
 }
 
 copyFiles
 
 # Rename i3 configs
-mv ~/.config/i3/config-${HOSTNAME} ~/.config/i3/config
-mv ~/.config/i3/i3status-${HOSTNAME}.conf ~/.config/i3/i3status.conf
+if [ -f "~/.config/i3/config-${HOSTNAME}" ]; then
+  mv ~/.config/i3/config-${HOSTNAME} ~/.config/i3/config
+  mv ~/.config/i3/i3status-${HOSTNAME}.conf ~/.config/i3/i3status.conf
+fi
